AccuracyCSS.cs file fully refactor(?).
Also fix some of the issues in the edge's dev tools.
Other issues are from the normalize css.

Move timer and start game click event listener to the html.

Remove the click event listener of the results tab that hides the tab.

Add an click event to close results button that hides the results tab.

Change the some variable names related to the timer.

Add comments to the timer.

havent finish refactoring AccuracyScript.js file.