Add a screenshot to the accuracy game project.
Put the color palette of matching game on css variables.

Move the right panel to the top and change its name.

Change the color palette of the mathcing game.

Add a title to matching game.

Install typescript using npm, set up the transpile, etc.

ts files and node_modules, js mapping are gitignored.

Refactor the code in  MatchingCSS.css file.

Refactor the code in MathcingScript.ts file (is not in the repository).

Add an image to the mathcing script project.