Fix matching misspeling.
Provide a description to the accuracy game and matching game.

Made the images in the matching game have a bit more quality

Add a smalles font-size to the page for mobile devices.

Made the project-box a grid element.

Made the project-box appear on top of media-mouseover-section.

Increase the font size of info-container as the element grows.