Decide to commit the ts files to the repository.
Add media links to the contact section.

Style chnages to the contact-section.

Made media-container show when on the about-section

Add new images to the accuracy and matching that show when hovering.

Add the 2 images for the flexidy project.

Add the 2 images for the social media simulator project.

Add the my resume.